Added help and example
Added text to include that the cmdlet does not return the current version of the file. Also added how to get that information. While I was in there I added another example because everyone likes more examples. Unless we're talking about Connect-PnPOnline, then maybe we'd like a few fewer examples. :)