Fix #3236: Directly operate on List
The old code retrieves the List object,
to then call a helper and pass the title of the list.

The helper, in turn, retrieves the list (again) by its title
to then add the ContentType.

Since the method already has a reference to the list,
we can forgo the helper.